Create README.md
# Bharatinternproject03
Titanic Classification is a machine learning task focused on predicting whether passengers on the Titanic survived or not based on various features such as age, gender, ticket class, and more. Using historical data, this model aims to learn patterns and relationships to accurately classify passengers into two categories: survivors and non-survivors. The outcome of this classification helps us understand factors that influenced survival rates during the tragic sinking of the Titanic.